requirements.txt
PyQt5==5.15.10
requests>=2.31.0
python-dotenv>=1.0.0
pyinstaller>=6.15.0

